requirements.txt
Django==5.0.6
djangorestframework==3.15.1
channels==4.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
daphne==4.1.2
djangorestframework-simplejwt==5.3.1
channels-redis==4.2.0
django-cors-headers==4.3.1
